fix: remove isDevelopment check and refactor code
