+++ src/misc.ts
@@ -536,9 +536,6 @@ export const combinedMultiAutocompleteChoices = (
   choices: MultiAutocompleteChoiceType[],
 ) => uniqBy([...selected, ...choices], "value");
 
-export const isInDevelopment =
-  !process.env.NODE_ENV || process.env.NODE_ENV === "development";
-
 export type WithOptional<T, K extends keyof T> = Omit<T, K> &
   Partial<Pick<T, K>>;
